Working on #1
This should now highlight missing fonts. This doesn't have complete accuracy and will need some testing on mac but should flag up some missing fonts. I'm currently investigating being able to pull in font files into the package to allow for a seamless transition to the target machines but this will take a bit more work for windows as it requires delving into the registry.